Save missing definitions when processing text
When a text was processed, the new occurrences where only added to the
template_occurrences model, so there was no way to conveniently retrieve the
occurrences lacking a definition.

Now listing words with an empty definition field will tell us there is an
existing occurrence that requires a meaning to be filled.